ml/training/requirements.txt
torch==2.8.0
torchvision==0.23.0
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
optuna>=3.0.0
wandb>=0.15.0
albumentations>=1.3.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorboard>=2.13.0
tensorflow>=2.13.0
onnx>=1.14.0
onnxruntime>=1.15.0
great-expectations>=0.17.0
dvc>=2.30.0
hydra-core>=1.3.0
omegaconf>=2.3.0
timm>=0.9.0
efficientnet-pytorch>=0.7.1
tensorboardX>=2.6
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jupyter>=1.0.0
ipywidgets>=8.0.0
